singly linked list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,5 +1,4 @@
 #include "lists.h"
-#include <stddef.h>
 #include <stdio.h>
 
 /**
@@ -11,7 +10,7 @@
 
 size_t print_list(const list_t *h)
 {
-	unsigned int index;
+	size_t index;
 
 	index = 0;
 	while (h != NULL)
